api/presents_service/requirements.txt
aiofiles==0.7.0
aiomysql==0.0.21
aiosqlite
alembic==1.6.5
appdirs==1.4.4
asgiref==3.4.1
asyncpg==0.23.0
bidict==0.21.2
black==21.7b0
click==8.0.1
colorama==0.4.4
databases==0.4.3
docstring-parser==0.7.1
ecdsa==0.17.0
fastapi==0.68.0
fastapi-crudrouter==0.8.0
fastapi-socketio==0.0.8
gino==1.0.1
gino-starlette==0.1.2
h11==0.12.0
httptools==0.2.0
inflection==0.5.1
iso8601==0.1.16
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
orm==0.1.6
ormar
pathspec==0.9.0
piccolo
psycopg2-binary==2.9.1
pyasn1==0.4.8
pydantic==1.8.2
PyMySQL==0.9.3
pypika-tortoise==0.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
python-engineio==4.2.1
python-jose==3.3.0
python-multipart==0.0.5
python-socketio==5.4.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.3.23
sqlalchemy-stubs==0.4
sqlparse==0.4.1
starlette==0.14.2
targ==0.3.5
toml==0.10.2
tomli==1.2.1
tortoise-orm==0.17.7
typesystem==0.2.4
typing-extensions==3.7.4.3
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
requests